# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/{organizationId}/sites/{siteId}/pages": {
      "get": {
        "description": "En el orden del sitio (`position`, siempre 0..n-1 sin huecos). Las de la papelera no aparecen. Solo pide membresía activa.",
        "operationId": "PagesController_list",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "siteId",
            "required": true,
            "in": "path",
            "description": "Sitio dueño de las páginas.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Páginas vivas del sitio, ordenadas.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "siteId": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "slug": {
                        "type": "string"
                      },
                      "position": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "visibility": {
                        "type": "string",
                        "enum": [
                          "PUBLIC",
                          "HIDDEN"
                        ]
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "DRAFT",
                          "PUBLISHED"
                        ]
                      },
                      "isHome": {
                        "type": "boolean"
                      },
                      "seoMeta": {
                        "nullable": true
                      },
                      "deletedAt": {
                        "nullable": true,
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      }
                    },
                    "required": [
                      "id",
                      "siteId",
                      "slug",
                      "position",
                      "visibility",
                      "status",
                      "isHome",
                      "seoMeta",
                      "deletedAt",
                      "createdAt",
                      "updatedAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (detalle en `issues`)."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "Sin membresía activa en la organización, sin el permiso requerido, o petición sin la cabecera anti-CSRF."
          },
          "404": {
            "description": "Sitio no encontrado, o de otra organización (ADR-002)."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Listar las páginas del sitio",
        "tags": [
          "pages"
        ]
      },
      "post": {
        "description": "Requiere el permiso `page.manage`. Se agrega al final del orden. `visibility` por defecto es `PUBLIC`; ocultarla la saca del menú pero sigue alcanzable por enlace directo.",
        "operationId": "PagesController_create",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "siteId",
            "required": true,
            "in": "path",
            "description": "Sitio donde se crea la página.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "slug": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 63,
                    "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
                  },
                  "visibility": {
                    "type": "string",
                    "enum": [
                      "PUBLIC",
                      "HIDDEN"
                    ]
                  }
                },
                "required": [
                  "slug"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Página creada en estado `DRAFT`, al final del orden.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "siteId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "position": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "visibility": {
                      "type": "string",
                      "enum": [
                        "PUBLIC",
                        "HIDDEN"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "DRAFT",
                        "PUBLISHED"
                      ]
                    },
                    "isHome": {
                      "type": "boolean"
                    },
                    "seoMeta": {
                      "nullable": true
                    },
                    "deletedAt": {
                      "nullable": true,
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    }
                  },
                  "required": [
                    "id",
                    "siteId",
                    "slug",
                    "position",
                    "visibility",
                    "status",
                    "isHome",
                    "seoMeta",
                    "deletedAt",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (detalle en `issues`)."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "Sin membresía activa en la organización, sin el permiso requerido, o petición sin la cabecera anti-CSRF."
          },
          "404": {
            "description": "Sitio no encontrado, o de otra organización (ADR-002)."
          },
          "409": {
            "description": "Ya existe una página con ese slug en este sitio."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Crear una página",
        "tags": [
          "pages"
        ]
      }
    },
...
    "/api/v1/organizations/{organizationId}/sites/{siteId}/pages/reorder": {
      "put": {
        "description": "Requiere `page.manage`. Se manda el orden **completo**, no «mové esta a la posición 3»: así el resultado no depende del orden en que lleguen varias peticiones ni deja huecos. La lista debe traer exactamente todas las páginas vivas del sitio, sin repetidos.",
        "operationId": "PagesController_reorder",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "siteId",
            "required": true,
            "in": "path",
            "description": "Sitio cuyas páginas se reordenan.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pageIds": {
                    "minItems": 1,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    }
                  }
                },
                "required": [
                  "pageIds"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Páginas en el nuevo orden.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "siteId": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "slug": {
                        "type": "string"
                      },
                      "position": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "visibility": {
                        "type": "string",
                        "enum": [
                          "PUBLIC",
                          "HIDDEN"
                        ]
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "DRAFT",
                          "PUBLISHED"
                        ]
                      },
                      "isHome": {
                        "type": "boolean"
                      },
                      "seoMeta": {
                        "nullable": true
                      },
                      "deletedAt": {
                        "nullable": true,
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      }
                    },
                    "required": [
                      "id",
                      "siteId",
                      "slug",
                      "position",
                      "visibility",
                      "status",
                      "isHome",
                      "seoMeta",
                      "deletedAt",
                      "createdAt",
                      "updatedAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "La lista tiene identificadores repetidos, o no incluye exactamente todas las páginas vivas del sitio."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "Sin membresía activa en la organización, sin el permiso requerido, o petición sin la cabecera anti-CSRF."
          },
          "404": {
            "description": "Sitio no encontrado, o de otra organización (ADR-002)."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Reordenar las páginas del sitio",
        "tags": [
          "pages"
        ]
      }
    },
...
  },
  "info": {
    "title": "Impulza One API",
    "description": "API REST de Impulza One. Todo cuelga de `/api/v1` salvo `/health`, que es un endpoint de\ninfraestructura sin sesión ni CSRF.\n\n**Autenticación**: cookie de sesión `HttpOnly` (`impulza_session`), emitida por\n`POST /auth/login`. No hay tokens en el cuerpo ni en cabeceras de autorización.\n\n**CSRF**: toda petición que modifica estado (POST, PATCH, PUT, DELETE) debe enviar la\ncabecera `X-Requested-With: impulza-one`. Un formulario o una imagen entre sitios no puede\nfijar cabeceras propias, y una petición `fetch` con cabecera custom dispara preflight CORS,\nbloqueado salvo origen autorizado. Sin ella la respuesta es 403.\n\n**Multi-tenant**: los recursos de negocio cuelgan de `/organizations/:organizationId/…`.\nUn recurso de otra organización responde **404 y no 403**, para no confirmar que existe\n(ADR-002).\n\n**Errores de validación**: 400 cuando el cuerpo no cumple el esquema de entrada, 422 cuando\ncumple la forma pero viola una regla de dominio (contraste insuficiente en un tema,\nconfiguración inválida para un tipo de bloque). Ambos traen `issues` con el campo exacto.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "sesion": {
        "type": "apiKey",
        "in": "cookie",
        "name": "impulza_session"
      }
    },
...
  }
}